requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3

django-ckeditor~=6.5.1
django-filter~=22.1
pytest~=7.2.0